Move SecretEditor into Components folder
